Las funciones de filtrado se encuentran ordenadas y colocadas en sus respectivos archivos
