+++ src/detalle.js
@@ -1,2 +1,2 @@
-const data = JSON.parse(localStorage.getItem("character"))
-console.log(data)
+//const data = JSON.parse(localStorage.getItem("character"))
+//console.log(data)
